% file metRology/man/apricot.Rd
% (c) 2011 S Ellison
\name{apricot}
\docType{data}
\alias{apricot}
\title{
Collaborative study results for fibre content in an apricot test material
}
\description{
A data frame containing reported duplicate results for dietary fibre from 
a collaborative study.
}
\usage{
	data(apricot)
}
\format{
A data frame containing duplicate results from 9 laboratories:
	\describe{
		\item{lab}{Factor giving abbreviated laboratory identifier}
		\item{fibre}{The reported fibre content.}
	}
}
\details{
Replicate results appear on separate rows.
}
\source{
B. W. Li, M. S. Cardozo (1994) Determination of total dietary fibre in foods 
and products with little or no starch, non-enzymatic gravimetric method: 
collaborative study.  
\emph{J. AOAC Int.} \bold{77}, 687-689, 1994
}
\references{
A. L. Ruhkin, C. J. Biggerstaff and M. G. Vangel (2000) Restricted maximum likelihood
estimation of a common mean and the Mandel-Paule algorithm. 
\emph{J. Stat. Planning an Inferences}
\bold{83}, 319-330, 2008

B. W. Li, M. S. Cardozo (1994) Determination of total dietary fibre in foods 
and products with little or no starch, non-enzymatic gravimetric method: 
collaborative study.  
\emph{J. AOAC Int.} \bold{77}, 687-689, 1994

}
\keyword{datasets}
